feat(pipeline): :sparkles: Fixing pipeline and mysql access failures
Starting the container in dettached mode allows the pipeline to execute the next command, that mostly will fail because mysql needs some time to iniatilize the first time.
Adding a logic to test the readiness of the mysql is essential to ensure the script can be run in the server after it completes its initialization.
This can be achieved using the mysqladmin command, and waiting until it connects succesfully.